new dashboard text
Addresses testing concern that the dashboard/homepage only says what to
do, not what this app is. I wanted to clarify that this is new (so it
replaces the long-hand written version), that it makes requirements
clear, and that it generates the completed doc. I think this lays out
the advantages for the user at this time.

The second paragraph is about what’s on the homepage/dashboard. I
didn’t call it a homepage or dashboard, though, since it doesn’t feel
like _obviously_ either of those things until there’s more capability.
What this page is should speak for itself while it’s still simple.