Allow 1550|postcss-custom-properties>...>url-regex dev-only dependency.
This is a development only dependency. `npm audit` is alerting on a Regular
Expression Denial of Service vulnerability within url-regex, as used by
postcss-custom-properties. See: npmjs.com/advisories/1550.

postcss-custom-properties is used only in development and CI to build our
frontend assets. This dependency is not used in production, staging, or
preview-deploys, which makes it impossible to perform a DoS attack via
this vulnerability.