Address frontend audit issues (#2440)
* Remove 1547|webpack>... items from allowlist, as suggested by audit-ci

* Allow 1548|webpack>...>serialize-javascript dev-only dependency.

This is a development only dependency. `npm audit` is alerting on a Remote Code
Execution vulnerability within serialize-javascript, as used by webpack.
See: npmjs.com/advisories/1548.

Webpack is only used in development and CI to build our frontend assets. This
dependency is not used in production, staging, or preview-deploys.

* Allow 1550|postcss-custom-properties>...>url-regex dev-only dependency.

This is a development only dependency. `npm audit` is alerting on a Regular
Expression Denial of Service vulnerability within url-regex, as used by
postcss-custom-properties. See: npmjs.com/advisories/1550.

postcss-custom-properties is used only in development and CI to build our
frontend assets. This dependency is not used in production, staging, or
preview-deploys, which makes it impossible to perform a DoS attack via
this vulnerability.